Scripts/requirements.txt
# Pharmacy Database System Requirements - Optimized for Render
# Using specific versions that build faster on Render

# Database
psycopg2-binary==2.9.7

# Web Framework
flask==2.3.3
flask-cors==4.0.0

# Environment Variables
python-dotenv==1.0.0

# PDF Processing - using a more stable version
PyMuPDF==1.23.26

# Data Processing - essential only
pandas==2.1.1

# Date/Time Handling
python-dateutil==2.8.2

# HTTP Requests
requests==2.31.0

# Scheduling
schedule==1.2.0

# Built-in modules (no installation needed):
# - imaplib (email processing)
# - email (email processing) 
# - json (data processing)
# - pathlib (file handling)
# - os (system operations)
# - datetime (date/time)
# - logging (logging)
# - threading (background tasks)
# - re (regex)
# - csv (CSV processing) 